Temporary fix: Revert to unsecured WiFi firmware for current build
The target is currently running an unsecured build but we configured it to use
secured firmware (.se files) which causes the MLAN driver to fail loading.

Changes:
- Revert WiFi firmware from .bin.se back to .bin
- Re-enable unsecured firmware package dependency

This is a temporary fix until the new AHAB secure boot build is deployed.
After the secure boot build is active, we'll switch back to .se firmware.